misc.process support windows

+++ misc/process.py
@@ -1,6 +1,9 @@
 # -*- coding: utf-8 -*-
 import os
 import signal
+import platform
+
+_system = platform.system().lower()
 
 __all__ = ['get_pid', 'process_start', 'process_stop', 'process_kill']
 
@@ -11,7 +14,11 @@ def get_pid(name):
     :param name: process name
     :return: pid list
     """
-    cmd = os.popen('pidof {0:s}'.format(name))
+    if _system == "windows":
+        cmd = os.popen("tasklist | grep {0:s} | awk '{{ print $2 }}'".format(name))
+    else:
+        cmd = os.popen('pidof {0:s}'.format(name))
+
     data = cmd.read().strip()
     if not data:
         return []
@@ -29,8 +36,13 @@ def process_control(name, cmd, start):
     retry = 3
     lst = list()
     assert isinstance(cmd, str), 'process cmd error!'
-    if not cmd.endswith('&'):
-        cmd += ' &'
+
+    # Make program running in background
+    if _system == "windows":
+        cmd = "start /b {0:s}".format(cmd)
+    else:
+        if not cmd.endswith('&'):
+            cmd += ' &'
 
     while not lst and retry > 0:
         lst = get_pid(name)
@@ -39,8 +51,9 @@ def process_control(name, cmd, start):
             retry -= 1
             continue
 
-        for pid in lst:
-            os.kill(pid, signal.SIGCONT if start else signal.SIGSTOP)
+        if _system != "windows":
+            for pid in lst:
+                os.kill(pid, signal.SIGCONT if start else signal.SIGSTOP)
 
 
 def process_start(name, cmd):
@@ -49,12 +62,15 @@ def process_start(name, cmd):
 
 
 def process_stop(name, cmd):
+    if _system == "windows":
+        return process_kill(name)
     process_control(name, cmd, False)
     return True if get_pid(name) else False
 
 
 def process_kill(name):
+    kill_signal = signal.SIGILL if _system == "windows" else signal.SIGKILL
     for pid in get_pid(name):
-        os.kill(pid, signal.SIGKILL)
+        os.kill(pid, kill_signal)
 
     return True if not get_pid(name) else False
